stm32hal: implement TX queue
When there's high-frequency data streams and high contention
on the bus, sending of synchronous PDOs might starve asynchronous
high-frequency data streams. The scenario is likely as follows:

- The node tries to send many synchronous PDOs putting them into a mailbox
- The node tries to send TPDO1 (high frequency) but the mailbox is full
  so it just sets the bufferFull flag
- The node tries to send TPDO1 again but it hasn't had a chance to
  transmit anything in between.
- Since the bufferFull flag is already set, it means we've already
  overwritten data pending for TX so we issue an error and go to PREOP

In order to avoid this, we implement a simple internal queue and
just buffer the data in there, also taking it from there whenever
a TX is successful.
Notice how when this mechanism is in place, we should remove the
bufferFull flag but it's also used somewhere else.
In order to enable this mechanism you must #define CO_USE_TX_QUEUE