TPDO: implement explicit sending based on variables
The current CoS (change of state) implementation is based on
the actual value of variables, by comparing the new value
with the previous one.
In some cases we would like to notify the CANopen stack to
explicitly retransmit all PDOs (if any) mapping a given
dictionary object, regardless of whether the value has actually
changed or not, without hardcoding the TPDO number.

Introduce two different approaches:

1) Simpler but slower. Just call CO_touch_TPDOvar() with
the address of the object being affected. This will look
up all TPDOs and mark for sending those affected.

2) Faster but more complex. Call once (at startup or after
the mapping has been changed) CO_get_TPDOmask() for all the
affected variables and make an arithmetic OR of all
the results. Save that somewhere. Call CO_touch_TPDOmask()
every time you need to trigger the transmission.