tpdo: fix TPDOs not being sent
when setting sendRequest from within interrupt context,
a race condition has been observed where this request
is not really honored if CO_process_TPDO() is being called
too frequently.
This is because the routine might accidentally overwrite
sendRequest with a 0 instead of an atomic or.
To fix this, we use a local variable within CO_TPDO_process()
instead of overwriting the sendRequest flag.

Notice there is still a slight chance of a race condition
where sendRequest is being set to 1 for multiple TPDOs at
once (e.g. 1,2,3,4) from interrupt context, while CO_TPDO_process()
is being called from within a loop.
This means that instead of 1,2,3,4 you get for instance 3,4,1,2,
which is in some sense still OK.